Minor updates to resize inputs
